Move Kafka and Google Cloud Pub/Sub sinks to be nested within a common package.
