+++ src/main/java/io/divolte/server/config/KafkaSinkConfiguration.java
@@ -27,8 +27,8 @@ import com.google.common.base.MoreObjects;
 
 import io.divolte.server.AvroRecordBuffer;
 import io.divolte.server.DivolteIdentifier;
-import io.divolte.server.kafka.KafkaFlushingPool;
-import io.divolte.server.kafka.Serializers;
+import io.divolte.server.topicsinks.kafka.KafkaFlushingPool;
+import io.divolte.server.topicsinks.kafka.Serializers;
 import org.apache.kafka.clients.producer.KafkaProducer;
 
 @ParametersAreNonnullByDefault
