+++ src/main/java/io/divolte/server/topicsinks/kafka/KafkaFlusher.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.divolte.server.kafka;
+package io.divolte.server.topicsinks.kafka;
 
 import com.google.common.collect.ImmutableList;
 import io.divolte.server.AvroRecordBuffer;
