+++ src/main/java/io/divolte/server/topicsinks/pubsub/GoogleCloudPubSubFlusher.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.divolte.server.pubsub;
+package io.divolte.server.topicsinks.pubsub;
 
 import com.google.api.core.ApiFuture;
 import com.google.api.gax.rpc.ApiException;
